Added check for chunk file size on disk. Will return a 500 if the size does not match.
