+++ server/views.py
@@ -2,23 +2,13 @@ import errno
 import os
 from functools import wraps
 
-from flask import jsonify, make_response, request, render_template
+from flask import request, render_template
 from server import app
 from server.models import File
 from server.utils import generate_auth_token, get_auth_status, get_auth_response, bam_test, \
     get_tempdir, get_chunk_filename, generate_file, remove_from_uploads, get_user_files, \
-    get_or_create_file, update_file_status, get_user_by_auth_token, InvalidServerToken
-
-
-def return_data(data, status_code=200):
-    if status_code >= 300:
-        app.logger.warning('{} - {}'.format(data, status_code))
-
-    return make_response(jsonify(data), status_code)
-
-
-def return_message(message, status_code):
-    return return_data({'message': message}, status_code=status_code)
+    get_or_create_file, update_file_status, get_user_by_auth_token, InvalidServerToken, \
+    return_data, return_message
 
 
 @app.errorhandler(500)
@@ -192,8 +182,9 @@ def chunk_upload(auth_token, sample_name, identifier, chunk_number):
 
     filename = request.form.get('flowFilename', type=str)
     total_chunks = request.form.get('flowTotalChunks', type=int)
+    current_chunk_size = request.form.get('flowCurrentChunkSize', type=int)
 
-    if not all([filename, total_chunks]):
+    if not all([filename, total_chunks, current_chunk_size]):
         return return_message('Error: missing parameter', 400)
 
     if not file or file.upload_status != 'ongoing':
@@ -218,6 +209,11 @@ def chunk_upload(auth_token, sample_name, identifier, chunk_number):
     except IOError:
         os.remove(chunk_filename)
         return return_message('Error: Chunk could not be saved', 500)
+
+    # Check that the chunk was saved in its integrity on the disk
+    if os.path.getsize(chunk_filename) != current_chunk_size:
+        os.remove(chunk_filename)
+        return return_message('Error: Chunk size on disk does not match', 500)
 
     # BAM test for integrity done here since only the prioritized last chunk is needed to determine corruption
     if chunk_number == total_chunks and file.file_type == 'BAM/SAM':
